Upgrade reservation feature & upload images

+++ resources/views/admin/event/list.blade.php
@@ -34,9 +34,9 @@
                                     <td>{{$ev->id}}</td>
                                     <td>{{$ev->name}}</td>
                                     <td>{{$ev->body}}</td>
-                                    <td>{{$ev->image}}</td>
+                                    <td> <img style="max-width: 200px;max-height: 200px;" src="{{$ev->image}}"> </td>
                                     <td class="center"><i class="fa fa-pencil fa-fw"></i> <a href="admin/event/edit/{{$ev->id}}">Edit</a></td>
-                                    <td class="center"><i class="fa fa-trash-o  fa-fw"></i><a href="admin/event/delete/{{$ev->id}}"> Delete</a></td>
+                                    <td class="center"><i class="fa fa-trash-o  fa-fw"></i><a href="admin/event/delete/{{$ev->id}}" onclick="return confirm('Bạn có chắc muốn xóa ?');"> Delete</a></td>
                                     
                                 </tr>
                             @endforeach
